CAMEL-18090: camel-main - Loading properties with profiles for prod/dev/test
